fix(desktop): preserve launch token across dotenv overrides
Move the current-main dotenv mechanics unchanged behind the public loader
facade. Capture the marked Desktop session credential before those mechanics
are imported, then restore that exact pair after every override-capable dotenv
layer. A dotenv file therefore cannot manufacture launcher authority for a
later layer, while unmarked shell exports and marker-only launches keep the
normal dotenv precedence.

Keep served-token adoption, readiness, and renderer recovery with their later
owners (#95890, #98939, #101094). This commit carries only the surviving
credential-source boundary from #76958.

Prior implementation seam: #66120 (ILO). Related implementation: #90780.
Related provenance: #40680, #54034, #73628.
